+++ app/functions.h
@@ -43,7 +43,7 @@ private:
 
 public:
   LocalIndex();
-  int size();
+  size_t size();
   void clear();
   uint32_t add_path(const std::string &path_to_insert);
   void add_words(std::unordered_set<std::string> &words_to_insert,
@@ -218,20 +218,20 @@ private:
   static bool is_config_loaded;
   static bool is_mapped;
   static bool first_time;
-  static int buffer_size;
+  static int64_t buffer_size;
   static std::filesystem::path index_path;
-  static int paths_size;
-  static int paths_size_buffer;
-  static int paths_count_size;
-  static int paths_count_size_buffer;
-  static int words_size;
-  static int words_size_buffer;
-  static int words_f_size;
-  static int words_f_size_buffer;
-  static int reversed_size;
-  static int reversed_size_buffer;
-  static int additional_size;
-  static int additional_size_buffer;
+  static int64_t paths_size;
+  static int64_t paths_size_buffer;
+  static int64_t paths_count_size;
+  static int64_t paths_count_size_buffer;
+  static int64_t words_size;
+  static int64_t words_size_buffer;
+  static int64_t words_f_size;
+  static int64_t words_f_size_buffer;
+  static int64_t reversed_size;
+  static int64_t reversed_size_buffer;
+  static int64_t additional_size;
+  static int64_t additional_size_buffer;
   static mio::mmap_sink mmap_paths;
   static mio::mmap_sink mmap_paths_count;
   static mio::mmap_sink mmap_words;
@@ -271,14 +271,14 @@ private:
       std::vector<Insertion> &to_insertions,
       int index_type); // index_type: 1 = words, 3 = reversed.
   static void write_to_transaction(std::vector<Transaction> &transactions,
-                        mio::mmap_sink &mmap_transactions,
-                        size_t &transaction_file_location);
+                                   mio::mmap_sink &mmap_transactions,
+                                   size_t &transaction_file_location);
   static int merge(index_combine_data &index_to_add);
   static int execute_transactions();
 
 public:
   static void save_config(const std::filesystem::path &config_index_path,
-                          const int config_buffer_size);
+                          const int64_t config_buffer_size);
   static bool is_index_mapped();
   static int initialize();
   static int uninitialize();
